requirements.txt
streamlit
sqlalchemy
pandas
psycopg2-binary
python-dotenv
pyyaml
pydantic